move impl to `formula_scorer.rs`
